requirements.txt


opencv-python
numpy
pandas
scikit-learn
streamlit
matplotlib
scipy
Pillow
